Properly update Credit Card form on add success
Changes the way state is handled in `EditPaymentMethods`. We don't duplicate the state anymore and use the parent state from `EditCollectiveForm` as the single source of thruth.

Now when the mutation query succeed, it returns the full payments list and local state is updated with the new values.

**Advantages**
- No more duplicated state
- We're sure that our list will always be up-to-date
- We get the PaymentMethod id in the list once added (and thus we can edit it)
- We don't need to refresh the page to add a new payment method anymore

**Drawbacks**
As we now fetch more data the query is slightly slower, but nothing's blocking.