Enforce donation amount to be an increment of 1
Enforce amount to be an increment of 1...
- On contribution custom amount input
- On tiers edit amount form

**About existing tiers with a floating amount**

In production, I've identified 12 tiers that are using floating amounts. Two of these tiers actually have backers.

Nothing will change for them, and people will still be able to subscribe to the tiers.

However if one of the admin tries to edit tier amount, he will not be able to set a floating value for the new amount.

TLDR ; No change for existing users, will enforce integer amount if updating tier.